Improve error handling in Mint.HTTP2.stream/2
Closes [TODO]

Before this commit, we were not correctly handling all errors from 
stream/2 and sometimes we would lose responses. We were sometimes 
throwing without responses and sometimes with responses from inside 
handle_data/3 and its nested calls, and catching in stream/2. Now 
instead, we're catching inside handle_data/3 so that we always re-throw 
with all the responses we have at that moment.