Refactor HTTP/2 tests with assert_recv_frames
This macro hides the call to recv_next_frames/1. The benefit of doing 
this is that we can infer the number of frames we expect by the length 
of the list of frames that we pass to the assert_recv_frames/1 macro. 
Plus, when we know that "assert record() = ..." works correctly, we can 
replace it in only one place (inside the new macro).