Check for client's max concurrent streams limit in HTTP/2 (#132)
Closes #119.

Before this commit, we didn't check for the client's max concurrent 
streams limit on the client side when HEADERS for a promised stream 
would come. Now we do that, and send an RST_STREAM with error code 
REFUSED_STREAM if the server tries to open a stream that goes over the 
client's limit.

We also do some more checks on the stream ID of a promised stream, as an 
additional security measure.