Fix a bug in the HTTP/2 handshake
In the HTTP/2 handshake, we expect the server to send a SETTINGS
frame right after connecting to it and sending it the connection
preface. However, the server can also send a GOAWAY and we were
not handling that. Now, we handle the server sending either
SETTINGS, GOAWAY, or any other frame with a more graceful error.