Fix bug with HTTP/2 requests exceeding window size (#161)
Before this commit, we had a bug when sending HTTP/2 requests with a
body exceeding the max window size of either the stream or the
connection. Basically, we were sending a HEADERS frame and only then
checking for the window size and possibly error out. This meant that
sending a request that exceeded the window size resulted in sending
the HEADERS and then erroring out, leaving the request in an open
state on the server.

With this commit, we build the payload for HEADERS and DATA without
sending it right away, and send them together if everything is
correct (which also leads to an optimization since we do only one
socket send call instead of two when sending requests with headers
and body).

I also fixed a small bug when checking the window size: before we
were checking if the size of the request body was greater or equal
to the window sizes, but the right check is just strictly greater
than the window size. Now, if you have a window size of 5 and send
a body of 5 bytes, it will work correctly.