Add support for trailing headers in HTTP/2
This change required to reword the way we handle incoming
"HEADERS" (plus possible "CONTINUATION") frames in HTTP/2.
Before this commit, we would decode the headers and then for
example check that the :status header was present.

Now, we need to track whether the first bunch of headers (the
*request* headers) were received for each stream. If they were
received, then the next headers that come for that stream have
to be trailing headers and have to have the END_STREAM flag set
(otherwise we do a protocol error).