Prevent open requests from leaking when HTTP2.request/5 returns an error
HTTP2 leaks open requests in some cases when there are errors raised in `HTTP2.request/5`.

I ran into this when hitting `:exceeds_window_size` and `:too_many_concurrent_requests` errors in production.
Over a long period of time we get some of these errors normally and each one leaks a request. When this happens, we retry on a different connection and the leak is not easily noticeable.
Unfortunately we eventually end up with enough leaked requests so that every request will error with `:too_many_concurrent_requests`.

This PR introduces some logic to clean up these errored requests, and adds a few checks to the tests to confirm the fix.