Make `docker-compose.yml` compatible with SELinux systems
While attempting to add a naive "Happy Eyeballs" option to test and experiment with, I encountered a "permission denied" error when running the Caddy container.

To resolve this issue, this commit adds the :z mount option, which ensures SELinux compatibility by addressing the permission issues caused by SELinux when the Caddyfile is mounted without the appropriate label. 

It's important to note that this change should only affect systems with SELinux enabled.