Optimize HTTP/2 request creation (#423)
When constructing the authority pseudo header we check if the conn's
port is the default for the conn's scheme using `URI.default_port/1`.
That corresponds to an ETS lookup into the `:elixir_config` table. It's
relatively fast to read that information but the conn's port and scheme
are static for the life of the conn, so we should determine this
information once while initiating the conn (`Mint.HTTP2.initiate/5`).
This change saves a small amount of time per request and becomes more
valuable as the conn is re-used for more requests.

Using a charlist is slightly faster than passing `"CONNECT"` as a
binary.