Add documentation note about protocol specific options
Most outside documenation (Finch, Req) links just to
`Mint.HTTP.connect/4` for documentation. That makes it somewhat
hard to find both the case insesntitive header options as well
as the not validating target options.

Not sure it's worth duplicating them there, so I figured I'd just
point to them - even if HTTP2 has no custom options as of now.

That's, of course, debatable.